３２ビットから６４ビットを起動したとき（あるいは逆）エラーになる問題を修正

+++ smartmv.cpp
@@ -164,7 +164,7 @@ bool tryAndArchive(LPCTSTR pFileOrig, LPCTSTR pRenameFull, LPCTSTR pRenamee)
 						{
 							CCommandLineString cls(cmdLine.c_str());
 							cls.remove(0); // remove exe
-							size_t toIndex = cls.getIndex(L"-to");
+							int toIndex = cls.getIndex(L"-to");
 							if (toIndex >= 0)
 								cls.remove(toIndex, 2);
 							arg = cls.toString();
